+++ dart/geobase/lib/src/vector/formats/wkb/wkb_decoder.dart
@@ -8,17 +8,24 @@ part of 'wkb_format.dart';
 
 class _WkbGeometryDecoder implements ContentDecoder {
   final GeometryContent builder;
-  final WkbConf conf;
 
-  _WkbGeometryDecoder(this.builder, {WkbConf? conf})
-      : conf = conf ?? const WkbConf();
+  /// When true, geometries decoded and detected as "empty" are built with
+  /// `emptyGeometry` method of content builder.
+  ///
+  /// If this is false (as it is by default) geometries detected as "empty" are
+  /// built with content methods specific to geometries. For example an empty
+  /// point (NaN, NaN) via `point` method and an empty line string (with 0
+  /// points) via `lineString` method.
+  final bool buildEmptyGeometries;
+
+  _WkbGeometryDecoder(this.builder, {this.buildEmptyGeometries = false});
 
   @override
   void decodeBytes(Uint8List source, {Map<String, dynamic>? options}) {
     _WkbGeometryBufferDecoder(
       builder,
       ByteReader.view(source),
-      conf,
+      buildEmptyGeometries: buildEmptyGeometries,
     ).buildAll();
   }
 
@@ -34,9 +41,13 @@ class _WkbGeometryDecoder implements ContentDecoder {
 class _WkbGeometryBufferDecoder {
   final GeometryContent builder;
   final ByteReader buffer;
-  final WkbConf conf;
+  final bool buildEmptyGeometries;
 
-  _WkbGeometryBufferDecoder(this.builder, this.buffer, this.conf);
+  _WkbGeometryBufferDecoder(
+    this.builder,
+    this.buffer, {
+    this.buildEmptyGeometries = false,
+  });
 
   void buildAll() {
     // loop as long as some data available, so builds all geometries from buffer
@@ -91,8 +102,11 @@ class _WkbGeometryBufferDecoder {
         builder.geometryCollection(
           count: numGeometries,
           // use callback content interface to build "numGeometries" to buffer
-          (geom) => _WkbGeometryBufferDecoder(geom, buffer, conf)
-              .buildCounted(numGeometries),
+          (geom) => _WkbGeometryBufferDecoder(
+            geom,
+            buffer,
+            buildEmptyGeometries: buildEmptyGeometries,
+          ).buildCounted(numGeometries),
         );
         break;
     }
@@ -100,7 +114,7 @@ class _WkbGeometryBufferDecoder {
 
   void _buildPoint(Coords coordType, Endian endian) {
     final point = _readPosition(coordType, endian);
-    if (conf.buildEmptyGeometries && point[0].isNaN && point[1].isNaN) {
+    if (buildEmptyGeometries && point[0].isNaN && point[1].isNaN) {
       // this is a special case, see => https://trac.osgeo.org/geos/ticket/1005
       //                             https://trac.osgeo.org/postgis/ticket/3181
       builder.emptyGeometry(Geom.point);
@@ -111,7 +125,7 @@ class _WkbGeometryBufferDecoder {
 
   void _buildLineString(Coords coordType, Endian endian) {
     final array = _readFlatPositionArray(coordType, endian);
-    if (conf.buildEmptyGeometries && array.isEmpty) {
+    if (buildEmptyGeometries && array.isEmpty) {
       builder.emptyGeometry(Geom.lineString);
     } else {
       builder.lineString(array, type: coordType);
@@ -120,7 +134,7 @@ class _WkbGeometryBufferDecoder {
 
   void _buildPolygon(Coords coordType, Endian endian) {
     final array = _readFlatLineStringArray(coordType, endian);
-    if (conf.buildEmptyGeometries && array.isEmpty) {
+    if (buildEmptyGeometries && array.isEmpty) {
       builder.emptyGeometry(Geom.polygon);
     } else {
       builder.polygon(array, type: coordType);
@@ -130,7 +144,7 @@ class _WkbGeometryBufferDecoder {
   void _buildMultiPoint(Coords coordType, Endian endian) {
     final array =
         _readPositionArray(coordType, endian, requireHeaderForItems: true);
-    if (conf.buildEmptyGeometries && array.isEmpty) {
+    if (buildEmptyGeometries && array.isEmpty) {
       builder.emptyGeometry(Geom.multiPoint);
     } else {
       builder.multiPoint(array, type: coordType);
@@ -143,7 +157,7 @@ class _WkbGeometryBufferDecoder {
       endian,
       requireHeaderForItems: true,
     );
-    if (conf.buildEmptyGeometries && array.isEmpty) {
+    if (buildEmptyGeometries && array.isEmpty) {
       builder.emptyGeometry(Geom.multiLineString);
     } else {
       builder.multiLineString(array, type: coordType);
@@ -153,7 +167,7 @@ class _WkbGeometryBufferDecoder {
   void _buildMultiPolygon(Coords coordType, Endian endian) {
     final array =
         _readFlatPolygonArray(coordType, endian, requireHeaderForItems: true);
-    if (conf.buildEmptyGeometries && array.isEmpty) {
+    if (buildEmptyGeometries && array.isEmpty) {
       builder.emptyGeometry(Geom.multiPolygon);
     } else {
       builder.multiPolygon(array, type: coordType);